+++ src/azure-cli-core/azure/cli/core/util.py
@@ -150,13 +150,13 @@ def handle_exception(ex):  # pylint: disable=too-many-locals, too-many-statement
         az_error = azclierror.CLIInternalError(ex.strerror)
         if ex.errno == 0:
             az_error.set_recommendation(
-                "Please report to us via Github: https://github.com/Azure/azure-cli/issues/20278")
+                "Please report to us via GitHub: https://github.com/Azure/azure-cli/issues/20278")
         elif ex.errno == -2146893813:
             az_error.set_recommendation(
-                "Please report to us via Github: https://github.com/Azure/azure-cli/issues/20231")
+                "Please report to us via GitHub: https://github.com/Azure/azure-cli/issues/20231")
         elif ex.errno == -2146892987:
             az_error.set_recommendation(
-                "Please report to us via Github: https://github.com/Azure/azure-cli/issues/21010")
+                "Please report to us via GitHub: https://github.com/Azure/azure-cli/issues/21010")
 
     else:
         error_msg = "The command failed with an unexpected error. Here is the traceback:"